requirements.txt
numpy==1.23.3
opencv_contrib_python==4.6.0.66
Pillow==9.4.0
pymongo==4.3.3
streamlit==1.20.0
#torch==1.13.0
torchvision==0.14.0
jinja2==3.1.2
